Add kubectl-ng config use-context
